requirements.txt
click==8.0.4
colorzero==1.1
distro==1.5.0
Flask~=2.0.2
gpiozero==1.6.2
idna==2.10
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mfrc522==0.0.7
python-mpd2==3.0.5
RPi.GPIO==0.7.0
spidev==3.5
Werkzeug~=2.0.2
python-mpd2==3.0.3